% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{ANOVA}
\alias{ANOVA}
\title{ANOVA}
\usage{
ANOVA(dataset, plotname = "", covariates = NULL)
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{plotname}{The name to be displayed on created plots}

\item{covariates}{A character vector of covariates to be included in the model. Example: c("Age","Sex")> Adding covariates will perform a Type III Multi-way ANOVA}
}
\value{
A list object containing the results of the ANOVA, the significant features and a volcano plot
}
\description{
This function performs differential expression analysis using an ANOVA.
}
