% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{assign_colors}
\alias{assign_colors}
\title{assign_colors}
\usage{
assign_colors(chars)
}
\arguments{
\item{chars}{A vector of unique entries (like unique(dataset$Status))}
}
\value{
A color mapping
}
\description{
This function assigns colors to entries in a vector
}
