% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{AUCs}
\alias{AUCs}
\title{AUCs}
\usage{
AUCs(dataset, PoIs, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{PoIs}{A vector containing the Proteins of interest. Example: c(""Q8TF72_SHROOM3" , "Q9ULZ3_PYCARD") or (unique(dataset$Protein))}

\item{plotname}{The name to be displayed on created plots}
}
\value{
A list object containing the results of the AUCs and a plot
}
\description{
Calculates the AUCs for a list of PoIs in the provided dataset.
}
