% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{BoxPlotsFeatures}
\alias{BoxPlotsFeatures}
\title{BoxPlotsFeatures}
\usage{
BoxPlotsFeatures(dataset, PoIs, plotname = "", pellet = "custom_vibrant")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{PoIs}{The vector containing the protein names}

\item{plotname}{The name to be displayed on created plots}
}
\value{
A plot object
}
\description{
Creates boxplots of any protein from a vector containing protein or peptide names.
}
