% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{ComBat}
\alias{ComBat}
\title{ComBat}
\usage{
ComBat(dataset, UseDiseaseCovariate = F)
}
\arguments{
\item{dataset}{The dataset to be corrected}

\item{UseDiseaseCovariate}{A boolean value indicating whether to use the disease status as a covariate in the ComBat correction}
}
\value{
The corrected dataset
}
\description{
This function corrects batch effects using ComBat
}
