% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{CompletenessAnalysis}
\alias{CompletenessAnalysis}
\title{CompletenessAnalysis}
\usage{
CompletenessAnalysis(dataset)
}
\arguments{
\item{dataset}{The dataset to be analyzed}
}
\value{
A list containing the results and a plot
}
\description{
This function calculates the number of unique proteins at different completeness levels
}
