% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{CorrelateFeatures}
\alias{CorrelateFeatures}
\title{CorrelateFeatures}
\usage{
CorrelateFeatures(
  dataset,
  vars1,
  vars2,
  use = "Protein",
  clustdist = "euclidean"
)
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{vars1}{The first list of features of interest}

\item{vars2}{The second list of features of interest}

\item{use}{The type of features to be used (Protein, Peptide, Gene)}

\item{clustdist}{The distance metric to be used for clustering}
}
\value{
A list object containing the correlation matrix, the correlation results and the heatmap
}
\description{
Correlation analysis for the specified dataset.
}
