% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{CorrelationNetwork}
\alias{CorrelationNetwork}
\title{CorrelationNetwork}
\usage{
CorrelationNetwork(dataset, cutoff = 0.7, cor.method = "pearson")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{cutoff}{The correlation cutoff to be used for the network (default is 0.7)}

\item{cor.method}{The correlation method to be used (default is "pearson")}
}
\value{
An interactive network plot
}
\description{
Correlation network for the specified dataset.
}
