% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{EffectAnalysis}
\alias{EffectAnalysis}
\title{EffectAnalysis}
\usage{
EffectAnalysis(dataset, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be tested}
}
\value{
A list object containing the results of the effect analysis
}
\description{
Calculates the correlation of clinical variables with principal components.
}
