% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{FindNACutoff}
\alias{FindNACutoff}
\title{FindNACutoff}
\usage{
FindNACutoff(dataset, plotname = "", p.adj.method = "BH", Test = "Wilcox")
}
\arguments{
\item{dataset}{The dataset to be filtered}

\item{plotname}{The name to be displayed on created plots}

\item{p.adj.method}{The method to be used for p-value adjustment ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{Test}{The test to be used for the analysis ("Wilcox", "Student")}
}
\description{
This function finds the optimal NA cutoff value to maximize the number of significant features by t-test
}
