% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{FisherTest}
\alias{FisherTest}
\title{FisherTest}
\usage{
FisherTest(dataset, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{plotname}{The name to be displayed on created plots}
}
\value{
A list object containing the results of the Fisher test, the significant features and a volcano plot
}
\description{
This function performs a Fisher test for differential expression analysis.
}
