% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{GenerateSampleIDsFromFilePath}
\alias{GenerateSampleIDsFromFilePath}
\title{GenerateSampleIDsFromFilePath}
\usage{
GenerateSampleIDsFromFilePath(filepath)
}
\arguments{
\item{filepath}{The path to the dataset}
}
\value{
A dataset with random sample IDs
}
\description{
This function generates random sample IDs for a given dataset
}
