% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{GLM}
\alias{GLM}
\title{GLM}
\usage{
GLM(dataset, PoIs, plotname = "", FalseNegativeWeight = 1, prevalence = "auto")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{PoIs}{A vector containing the Proteins of interest. Example: c(""Q8TF72_SHROOM3" , "Q9ULZ3_PYCARD") or (unique(dataset$Protein))}

\item{plotname}{The name to be displayed on created plots}

\item{FalseNegativeWeight}{The relative cost assigned to false negatives when calculating Youden's index (lower values will prioritize sensitivity, higher values will prioritize specificity). Recommended Values between 0.2 and 5}

\item{crossvalidation}{A boolean value indicating whether to use cross-validation> I recommend false for desriptive analysese and true for predictife tasks}
}
\value{
A list object containing the results of the GLM model, the confusion matrix and the ROC plot and the AUC if the dataset has 2 classes
}
\description{
Generalized linear model for the specified dataset.
}
