% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{GSEA}
\alias{GSEA}
\title{GSEA}
\usage{
GSEA(GeneList, plotname = "", collections = c("C2", "C5"))
}
\arguments{
\item{GeneList}{An ordered named vector of gene-level statistics (e.g., log fold changes) where names are gene symbols.}

\item{plotname}{The name to be displayed on created plots}

\item{collections}{A character vector specifying the collections to be used for GSEA (e.g., c("C2", "C5")). Default is c("C2", "C5").}
}
\value{
A list object containing the results of the GSEA analysis, the GSEA results table, and the GSEA enrichment plots.
}
\description{
Gene set enrichment analysis for the specified dataset.
}
