% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{HeatMap}
\alias{HeatMap}
\title{HeatMap}
\usage{
HeatMap(
  dataset,
  PoIs,
  method = "unsupervised",
  clustDist = "euclidean",
  plotname = "",
  show_column_names = FALSE,
  show_row_names = FALSE,
  column_split = NULL,
  row_split = NULL,
  cluster_columns = TRUE,
  Annotations = NULL,
  ColPalette = "Set1",
  contColors = c("yellow", "purple"),
  fontsize = 8
)
}
\arguments{
\item{dataset}{The dataset to be plottet}

\item{PoIs}{The list of proteins or peptides of interest}

\item{method}{The method to be used for the Heatmap (unsupervised, supervised)}

\item{clustDist}{The distance metric to be used for clustering in the Heatmap ("euclidean", "maximum", "man-hattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall")}

\item{plotname}{The name to be displayed on created plots}

\item{show_column_names}{Logical value indicating if column names should be shown}

\item{show_row_names}{Logical value indicating if row names should be shown}

\item{column_split}{A numeric indicating into how many clusters the columns should be split}

\item{row_split}{A numeric indicating into how many clusters the rows should be split}

\item{cluster_columns}{Logical value indicating if columns should be clustered at all}

\item{Annotations}{comumn names of the dataset that should be annotated}

\item{ColPalette}{The color palette to be used for the annotations (Specified using the names of a color palette from the RColorBrewer package): https://www.aptech.com/releases/gauss18/graphics-updates/color-brewer-palettes/}

\item{contColors}{The colors to be used for continuous annotations of numeric columns (Input is a vector of colors, example c("green", "red"))}

\item{fontsize}{The fontsize to be used for the annotations (numeric value)}
}
\value{
A Heatmap object
}
\description{
veratile Heatmap function that can be used for unsupervised and supervised clustering of a list of proteins or peptides
}
