% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{HeatMapProteinClusterAnalysis}
\alias{HeatMapProteinClusterAnalysis}
\title{HeatMapProteinClusterAnalysis}
\usage{
HeatMapProteinClusterAnalysis(HeatMap, STRINGBackground, n)
}
\arguments{
\item{HeatMap}{The heatmap object to be analyzed}

\item{n}{The number of Protein lusters to be created}
}
\value{
A list object containing the proteins in each cluster, the original heatmap, the modified heatmap and the gene ontology analysis for each cluster
}
\description{
Extension of the Heatmap function. This function takes a heatmap object and clusters the proteins into n clusters. It then performs gene ontology analysis on each cluster.
}
