% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{ImportMSData}
\alias{ImportMSData}
\title{ImportMSData}
\usage{
ImportMSData(filepath, programm, SampleID = FALSE, feature = "Protein")
}
\arguments{
\item{filepath}{The path to the dataset}

\item{programm}{The program used to generate the data (for now only DIA-NN is supported)}

\item{SampleID}{A boolean value indicating whether the data has sample IDs generated by the function GenerateSampleIDsFromFilePath}

\item{feature}{The feature to be imported (Protein or Peptide)}
}
\value{
A dataset with the imported data
}
\description{
This function imports MS data from DIA-NN output
}
