% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{kffs}
\alias{kffs}
\title{kffs}
\usage{
kffs(dataset, method = "SVM", kfolds = 10)
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{method}{The method to be used for feature selection (SVM, RF, LDA, GLM)}

\item{kfolds}{The number of folds to be used for feature selection}
}
\value{
A list object containing the results of the feature selection
}
\description{
k fold feature selection for the specified dataset.
}
