% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{KNN}
\alias{KNN}
\title{KNN}
\usage{
KNN(dataset, PoIs)
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{PoIs}{A vector containing the Proteins of interest. Example: c(""Q8TF72_SHROOM3" , "Q9ULZ3_PYCARD") or (unique(dataset$Protein))}
}
\value{
A list object containing the results of the KNN model, the confusion matrix and the ROC plot and the AUC if the dataset has 2 classes
}
\description{
K-Nearest Neighbors calssification for the specified dataset.
}
