% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{KruskalTest}
\alias{KruskalTest}
\title{KruskalTest}
\usage{
KruskalTest(
  dataset,
  plotname = "",
  clustDist = "euclidean",
  method = "unsupervised",
  p.adj.method = "BH"
)
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{plotname}{The name to be displayed on created plots}

\item{clustDist}{The distance metric to be used for clustering in the Heatmap ("euclidean", "maximum", "man-hattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall")}

\item{method}{The method to be used for the Heatmap (unsupervised, supervised)
#' @param p.adj.method The method to be used for p-value adjustment ("BH", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}
}
\value{
A list object containing the results of the Kruskal test, the significant features and a heatmap
}
\description{
This function performs differential expression analysis using a Kruskal test.
}
