% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{LASSO}
\alias{LASSO}
\title{LASSO}
\usage{
LASSO(
  dataset,
  PoIs,
  nCV = 10,
  plotname = "",
  standardize = T,
  NProt = NULL,
  covariates = NULL
)
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{PoIs}{A character vector of proteins of interest to be included in the L
ASSO model. Example: c("Protein1", "Protein2", "Protein3"). It is recommended to select proteins that are biologically relevant or have shown some association with the outcome variable in previous analyses.}

\item{nCV}{The number of cross-validation folds to be used for selecting the optimal}

\item{standardize}{A logical value indicating whether to standardize the features before fitting the model. Default is TRUE, which is recommended for LASSO regression to ensure that all features are on the same scale.}

\item{NProt}{An optional integer specifying the desired number of non-zero coefficients (selected proteins) in the final model. If NULL (default), the function will use the lambda value that minimizes cross-validated error.}

\item{covariates}{A character vector of covariates to be included in the model. Example: c("Age","Sex").}
}
\value{
A list object containing the fitted LASSO model, the best lambda value, the cross-validated AUC, and the selected proteins with non-zero coefficients.
}
\description{
Perform LASSO regression to identify a subset of proteins that are most predictive of the outcome variable (Status). LASSO is a regularization technique that can help prevent overfitting and improve model interpretability by shrinking some coefficients to zero, effectively performing feature selection.
}
