% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{LIMMA}
\alias{LIMMA}
\title{LIMMA}
\usage{
LIMMA(dataset, subject_var = NULL, covariates = NULL, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{subject_var}{The name of the variable in the dataset that identifies the subjects}

\item{covariates}{A character vector of covariates to be included in the model. Example: c("Age","Sex", TimePoint). It is recommended to order the levels of time series data.}

\item{plotname}{The name to be displayed on created plots}
}
\value{
A list object containing the results of the LIMMA analysis, the significant features and a volcano plot
}
\description{
This function performs differential expression analysis using the LIMMA package, which is designed
for analyzing complex experiments. It uses linear models to assess differential expression while accounting for covariates and multiple testing, for example repeated measurements from the same person.
}
