% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{LogisticRegressionSingleFeature}
\alias{LogisticRegressionSingleFeature}
\title{LogisticRegressionSingleFeature}
\usage{
LogisticRegressionSingleFeature(dataset, PoI)
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{PoI}{The feature of interest}
}
\value{
A plot object
}
\description{
Plots the logistic regression of a single feature.
}
