% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{LTest}
\alias{LTest}
\title{LTest}
\usage{
LTest(dataset, plotname = "", Covariates = NULL)
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{plotname}{The name to be displayed on created plots}

\item{Covariates}{A character vector of covariates to be included in the model. Example: c("Age","Sex")}
}
\value{
A list object containing the results of the logistic regression, the significant features and a volcano plot
}
\description{
This function performs differential expression analysis using a logistic regression model.
}
