% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{MultiWayComaprison}
\alias{MultiWayComaprison}
\title{MultiWayComaprison}
\usage{
MultiWayComaprison(dataset, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{plotname}{The name to be displayed on the resulting plots}
}
\value{
A list object containing the results of the Kruskal test and the PCA, as well a UMAP analysis
}
\description{
Multi way comparison of the specified dataset.
}
