% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{NaCutoff}
\alias{NaCutoff}
\title{NaCutoff}
\usage{
NaCutoff(dataset, cutoffvalue = 70)
}
\arguments{
\item{dataset}{The dataset to be filtered}

\item{cutoffvalue}{The cutoff value for the percentage of missing values}
}
\value{
A dataset with the filtered proteins
}
\description{
This pipeline friendly function gets rid of every protein that has more than x \% missing values (default = 70)
}
