% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{nObsPerFeature}
\alias{nObsPerFeature}
\title{nObsPerFeature}
\usage{
nObsPerFeature(dataset, n = 10)
}
\arguments{
\item{dataset}{The dataset to be filtered}

\item{n}{The number of observations}
}
\value{
The filtered dataset
}
\description{
This pipeline friendly function filters out every Protein with less than n observations total
}
