% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{normalizeIntensityOnFeature}
\alias{normalizeIntensityOnFeature}
\title{normalizeIntensityOnFeature}
\usage{
normalizeIntensityOnFeature(dataset, method = "minmax", plot = FALSE)
}
\arguments{
\item{dataset}{The dataset to be normalized}

\item{method}{The method to be used for normalization (mean, median, z-score, minmax)}

\item{plot}{A boolean value indicating whether a plot should be generated}
}
\value{
The normalized dataset
}
\description{
This function normalizes MS Intensities on a feature level
}
