% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{normalizeIntensityOnSample}
\alias{normalizeIntensityOnSample}
\title{normalizeIntensityOnSample}
\usage{
normalizeIntensityOnSample(dataset, method = "minmax", plot = FALSE)
}
\arguments{
\item{dataset}{The dataset to be normalized}

\item{method}{The method to be used for normalization (mean, median, z-score, minmax)}

\item{plot}{A boolean value indicating whether a plot should be generated}
}
\value{
The normalized dataset
}
\description{
This pipeline friendly function normalizes MS Intensities on a sample level
}
