% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{PlotHistogram}
\alias{PlotHistogram}
\title{PlotHistogram}
\usage{
PlotHistogram(dataset, PoIs, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{PoIs}{The protein or peptide of interest}

\item{plotname}{The name to be displayed on created plots}
}
\value{
A plot object
}
\description{
Creates a histogram of the intensity of a given protein.
}
