% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{PLSDA}
\alias{PLSDA}
\title{PLSDA}
\usage{
PLSDA(dataset, plotname = "", PoIs)
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{plotname}{The name to be displayed on created plots}

\item{PoIs}{A vector containing the Proteins of interest. Example: c(""Q8TF72_SHROOM3" , "Q9ULZ3_PYCARD") or (unique(dataset$Protein))}

\item{ncomp}{The number of components to be calculated}
}
\value{
A list object containing the results of the PLSDA model, 2D and 3D score and loading plots and the ROC plot if the dataset has 2 classes
}
\description{
Partial least squares discriminant analysis for the specified dataset.
}
