% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{PredictGLM}
\alias{PredictGLM}
\title{PredictGLM}
\usage{
PredictGLM(
  dataset,
  PoIs,
  Covariates = NULL,
  plotname = "",
  allowInteractions = FALSE
)
}
\arguments{
\item{dataset}{The dataset to be tested, It is recommended that Status is a factor with the controls as the first level}

\item{PoIs}{A vector containing the Proteins of interest.}

\item{Covariates}{A vector containing the covariates to be included in the model. Example: c("Age", "Sex")}

\item{plotname}{The name to be displayed on created plots}

\item{allowInteractions}{A logical value indicating whether to allow interactions between PoIs in the model. Default is FALSE}
}
\value{
A list object containing the results of the GLM predictions and a plot including ROC curve and calibration plot
}
\description{
Predicts the GLM for a list of PoIs in the specified dataset
}
