% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{RankedIntensities}
\alias{RankedIntensities}
\title{RankedIntensities}
\usage{
RankedIntensities(dataset, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be plotted}

\item{plotname}{The name to be displayed on created plots}
}
\value{
A list containing the log2 and normal plot and a table with the ranked intensities
}
\description{
This function plots the ranked mean intensities for each Protein
}
