% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{RemoveOutliers}
\alias{RemoveOutliers}
\title{RemoveOutliers}
\usage{
RemoveOutliers(dataset, Stdev = 2, plotname = "Outlier plot", plot = TRUE)
}
\arguments{
\item{dataset}{The dataset to be filtered}

\item{Stdev}{The standard deviation for the outlier removal}

\item{plotname}{The name to be displayed on created plots}

\item{plot}{A boolean value indicating whether a plot should be generated}
}
\value{
A list containing the outlier plot, a table with the outliers and the dataset without the outliers
}
\description{
This function removes outliers from the dataset
}
