% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{RemoveOutliers2}
\alias{RemoveOutliers2}
\title{RemoveOutliers2}
\usage{
RemoveOutliers2(dataset, StdDevs = 3, plotname = "Outlier Plot")
}
\arguments{
\item{dataset}{The dataset to be filtered}

\item{StdDevs}{The number of standard deviations to be used for outlier removal}
}
\value{
A list containing the dataset without outliers, a plot and a table with the number of unique proteins per sample
}
\description{
This function removes outliers from the dataset based on the number of detected proteins
}
