% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{ROC}
\alias{ROC}
\title{ROC}
\usage{
ROC(dataset, PoI, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{PoI}{The feature of interest}

\item{nCross}{The number of iterations for the cross validation}
}
\value{
A list object containing the results of the ROC calculations and a plot
}
\description{
Plots the ROC for a single feature in the specified dataset.
}
