% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{SampleRandomization}
\alias{SampleRandomization}
\title{SampleRandomization}
\usage{
SampleRandomization(df, group_vars, IDColumn)
}
\arguments{
\item{df}{The dataset to be randomized}

\item{group_vars}{The grouping variables}

\item{IDColumn}{The column containing the sample IDs}
}
\value{
A list containing the plate layout, summary statistics, and plots of the group_vars
}
\description{
This function designes the layout of a 96 well plate given a dataset and grouping variables whos distribution needs to be stable across plates
}
