% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{STRING}
\alias{STRING}
\title{STRING}
\usage{
STRING(PoIs, STRINGBackground, plotname = "", colPellet = "Reds")
}
\arguments{
\item{PoIs}{A vector containing the Proteins of interest. Example: c("Q8TF72_SHROOM3" , "Q9ULZ3_PYCARD")}

\item{plotname}{The name to be displayed on created plots}

\item{BackgroundSet}{A vector containing the background Proteins. Example: c(""Q8TF72_SHROOM3" , "Q9ULZ3_PYCARD") or (unique(dataset$Protein))}
}
\value{
A list object containing the results of the STRING analysis, the STRING network plot and the STRING enrichment plots
}
\description{
STRING analysis for the specified dataset.
}
