% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{TrajectoryAnalysis}
\alias{TrajectoryAnalysis}
\title{TrajectoryAnalysis}
\usage{
TrajectoryAnalysis(dataset, timecol, k = k, deepSplit = 3, minClusterSize = 50)
}
\arguments{
\item{dataset}{The dataset to be analyzed. Must contain columns "Protein", "Intensity", and a time column.}

\item{timecol}{The name of the time column in the dataset. Input as a string.}

\item{k}{The number of knots for the GAM spline. Default is 5.}

\item{deepSplit}{The deepSplit parameter for dynamic tree cutting. Default is 3}

\item{minClusterSize}{The minimum cluster size for dynamic tree cutting. Default is 50.}
}
\value{
A list object containing the GAM models, the predicted dynamics, the derivatives,
the cluster assignments, the mean derivative trajectories for each cluster, and plots of the Eigentrajectories for each cluster.
}
\description{
Trajectory analysis of the specified dataset. Identifies clusters of proteins with similar temporal dynamics based on their derivatives.
}
