% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{TwoWayComaprison}
\alias{TwoWayComaprison}
\title{TwoWayComaprison}
\usage{
TwoWayComaprison(dataset, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{plotname}{The name to be displayed on the resulting plots}
}
\value{
A list object containing the results of the Wilcox test, the PCA as well a UMAP analysis and the STRING analysis if there are significant results
}
\description{
Two way comparison of the specified dataset.
}
