% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{UMAP}
\alias{UMAP}
\title{UMAP}
\usage{
UMAP(dataset, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be plotted}

\item{plotname}{The name to be displayed on created plots}
}
\value{
A list object containing the results of the UMAP calculations and the UMAP plot
}
\description{
Uniform Manifold Approximation and Projection for the specified dataset.
}
