% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{WTest}
\alias{WTest}
\title{WTest}
\usage{
WTest(dataset, plotname = "", method = "unsupervised", clustDist = "euclidean")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{plotname}{The name to be displayed on created plots}

\item{method}{The method to be used for the Heatmap (unsupervised, supervised)}

\item{clustDist}{The distance metric to be used for clustering in the Heatmap ("euclidean", "maximum", "man-hattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall")}
}
\value{
A list object containing the results of the Wilcox test, the significant features and a volcano plot
}
\description{
This function performs differential expression analysis using a Wilcox test.
}
