% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{ANOVA}
\alias{ANOVA}
\title{ANOVA}
\usage{
ANOVA(dataset, plotname = "", clustDist = "euclidean", method = "unsupervised")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{plotname}{The name to be displayed on created plots}

\item{clustDist}{The distance metric to be used for clustering in the Heatmap ("euclidean", "maximum", "man-hattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall")}

\item{method}{The method to be used for the Heatmap (unsupervised, supervised)}
}
\value{
A list object containing the results of the ANOVA, the significant features and a heatmap
}
\description{
This function performs differential expression analysis using an ANOVA.
}
