% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{assign_colors}
\alias{assign_colors}
\title{assign_colors}
\usage{
assign_colors(unique_entries, palette = "Set1")
}
\arguments{
\item{unique_entries}{A vector of unique entries (like unique(dataset$Status))}

\item{palette}{The name of the Brewer palette to be used (default = "Set1")}
}
\value{
A color mapping
}
\description{
This function assigns colors from a Brewer’s palette to entries in a vector
}
