% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{BiomarkerPanel}
\alias{BiomarkerPanel}
\title{BiomarkerPanel}
\usage{
BiomarkerPanel(dataset, PoIs, n, FalseNegativeWeight = 1, prevalence = "auto")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{PoIs}{A vector containing the Proteins of interest}

\item{n}{The maximum number of PoIs to be included in the Biomarker Panel}

\item{FalseNegativeWeight}{The weight of false negatives in the ROC analysis (default is 1, recommended is 0.2 to 5)}

\item{prevalence}{The prevalence of the disease in the population. Used to bias the discovery towards higher specificity (default is "auto")}
}
\value{
A list object containing the results of the Biomarker Panel analysis
}
\description{
Biomarker panel for the specified dataset.
}
