% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{BiomarkerPanel}
\alias{BiomarkerPanel}
\title{BiomarkerPanel}
\usage{
BiomarkerPanel(dataset, PoIs, n, crossvalidation = F, p.adj.method = "BH")
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{PoIs}{A vector containing the Proteins of interest}

\item{n}{The maximum number of PoIs to be included in the Biomarker Panel}
}
\value{
A list object containing the results of the Biomarker Panel analysis
}
\description{
Biomarker panel for the specified dataset.
}
