% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{ComBat}
\alias{ComBat}
\title{ComBat}
\usage{
ComBat(dataset)
}
\arguments{
\item{dataset}{The dataset to be corrected}
}
\value{
The corrected dataset
}
\description{
This function corrects batch effects using ComBat
}
