% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{FindNACutoff}
\alias{FindNACutoff}
\title{FindNACutoff}
\usage{
FindNACutoff(dataset, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be filtered}

\item{plotname}{The name to be displayed on created plots}
}
\value{
A list object the dataset with the optimal NA cutoff value and a plot of the results
}
\description{
This function finds the optimal NA cutoff value to maximize the number of significant features by t-test
}
