% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{FisherTest}
\alias{FisherTest}
\title{FisherTest}
\usage{
FisherTest(dataset, p.adjust.method = "BH")
}
\arguments{
\item{dataset}{The dataset to be tested}
}
\value{
A list object containing the results of the Fisher test
}
\description{
This function performs a Fisher test for differential expression analysis.
}
