% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{ImportMSData}
\alias{ImportMSData}
\title{ImportMSData}
\usage{
ImportMSData(filepath, feature = "Protein", programm = "DIA-NN")
}
\arguments{
\item{filepath}{The path to the dataset}

\item{feature}{The feature to be imported (Protein or Peptide)}

\item{programm}{The program used to generate the data c("DIA-NN, "Spectronaut)}
}
\value{
A dataset with the imported data
}
\description{
This function imports MS data from DIA-NN output
}
