% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{ImputeFeatureIntensity}
\alias{ImputeFeatureIntensity}
\title{ImputeFeatureIntensity}
\usage{
ImputeFeatureIntensity(dataset, method = "half_min")
}
\arguments{
\item{dataset}{The dataset to be imputed}

\item{method}{The method to be used for imputation (mean, zero, half_min, median, knn)}
}
\value{
The imputed dataset
}
\description{
This pipeline friendly function imputes missing values in the dataset
}
