% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{LASSO}
\alias{LASSO}
\title{LASSO}
\usage{
LASSO(dataset, PoIs, nCV = 3)
}
\arguments{
\item{dataset}{The dataset to be tested}

\item{PoIs}{A character vector of proteins of interest to be included in the L
ASSO model. Example: c("Protein1", "Protein2", "Protein3"). It is recommended to select proteins that are biologically relevant or have shown some association with the outcome variable in previous analyses.}

\item{nCV}{The number of cross-validation folds to be used for selecting the optimal}
}
\value{
A list object containing the fitted LASSO model, the best lambda value, the cross-validated AUC, and the selected proteins with non-zero coefficients.
}
\description{
Perform LASSO regression to identify a subset of proteins that are most predictive of the outcome variable (Status). LASSO is a regularization technique that can help prevent overfitting and improve model interpretability by shrinking some coefficients to zero, effectively performing feature selection.
}
