% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{MEGENA}
\alias{MEGENA}
\title{MEGENA}
\usage{
MEGENA(dataset, plotname = "")
}
\arguments{
\item{dataset}{The dataset to be tested}
}
\value{
A list object containing the results of the MEGENA analysis, the module table, the hierarchy plot, the STRING analysis and the correlation plot
}
\description{
Multiscale Embedded Gene Co-expression Network Analysis (MEGENA) for the specified dataset.
}
